Path tracing!
Purely path traced using cosine-weighted ray bouncing. Image is very noisy, as to be expected I guess. I don't use the roughness to influence ray bouncing so that might be next. Kind of wanted purely path tracing to determine colour but might have to "cheat" so it isn't so noisy. Also uses a thread pool now.