Rendered and add images of a Cornell box
Not perfect, cubes aren't in the exact same place and the light isn't the same yellow colour. They describe the light in wavelengths instead of just a vec3. Shadows on the small cube look very similar to me and the light bleed from the green onto the floor look good too. My image is still a bit noisy, that was after 6000 frames (each pixel traced 6000 times), would be less noisy with a back wall so the pixels are more likely to bounce to a light eventually. Add support for models next.