Added editing of material on mesh
Can now change the individual materials of a mesh. Want to look into guided ray sampling, where we sometimes try pointing the ray in the direction of a light to help avoid when some pixels just happen to never hit a light. Could be complicated with materials in a mesh being emitters. May first of all just add that without the optimisation for mesh emitters. Mesh emitters would need to identify where the triangles that emit are I suppose and then we could randomly choose between them.